overhauled BasicOp subclasses
Constructors no longer take an owner, instead relying on the QueueBuilder to
set an owner for them. This will make it easier to deal with these classes
directly in the future.

Additionally, the old completion callback block has been removed from the
constructors, replaced with #before and #after methods which will register a
block to be called when the operation starts and finishes, respectively. This
makes the exec and wait operations more consistent with the others.